+++ src/cow-react/common/pure/CurrencyArrowSeparator/index.tsx
