+++ src/cow-react/common/pure/CurrencyArrowSeparator/style-mixins.tsx
