+++ src/cow-react/common/pure/CurrencyArrowSeparator/styled.tsx
