+++ src/cow-react/common/pure/CurrencyInputPanel/styled.tsx
