+++ src/cow-react/common/pure/CurrencyInputPanel/typings.ts
