+++ src/cow-react/modules/governance/pure/GovernanceList/index.cosmos.tsx
