+++ src/cow-react/modules/limitOrders/containers/LimitOrdersWidget/index.tsx
@@ -1,12 +1,12 @@
 import * as styledEl from './styled'
-import { CurrencyInputPanel } from 'cow-react/common/dumb/CurrencyInputPanel'
-import { CurrencyArrowSeparator } from 'cow-react/common/dumb/CurrencyArrowSeparator'
-import { AddRecipient } from 'cow-react/common/dumb/AddRecipient'
+import { CurrencyInputPanel } from 'cow-react/common/pure/CurrencyInputPanel'
+import { CurrencyArrowSeparator } from 'cow-react/common/pure/CurrencyArrowSeparator'
+import { AddRecipient } from 'cow-react/common/pure/AddRecipient'
 import { ButtonPrimary } from 'components/Button'
 import { Trans } from '@lingui/macro'
 import React from 'react'
 import { BalanceAndSubsidy } from 'hooks/useCowBalanceAndSubsidy'
-import { CurrencyInfo } from 'cow-react/common/dumb/CurrencyInputPanel/typings'
+import { CurrencyInfo } from 'cow-react/common/pure/CurrencyInputPanel/typings'
 import { Field } from 'state/swap/actions'
 import { GNO, USDC } from 'constants/tokens'
 import { ButtonSize } from 'theme'
