+++ src/cow-react/modules/swap/hooks/useSwapButtonContext.ts
@@ -17,7 +17,7 @@ import { useGnosisSafeInfo } from 'hooks/useGnosisSafeInfo'
 import { useIsSwapUnsupported } from 'hooks/useIsSwapUnsupported'
 import { ApproveButtonProps } from 'cow-react/modules/swap/smart/ApproveButton'
 import { getSwapButtonState } from 'cow-react/modules/swap/helpers/getSwapButtonState'
-import { SwapButtonContext } from 'cow-react/modules/swap/dumb/SwapButton/SwapButton'
+import { SwapButtonContext } from 'cow-react/modules/swap/pure/SwapButton/SwapButton'
 import { useGetQuoteAndStatus } from 'state/price/hooks'
 import { OperationType } from 'components/TransactionConfirmationModal'
 import { useTransactionConfirmModal } from 'cow-react/modules/swap/hooks/useTransactionConfirmModal'
