+++ src/cow-react/modules/swap/pure/CompatibilityIssuesWarning/index.tsx
