+++ src/cow-react/modules/swap/pure/ReceiveAmount/styled.tsx
