+++ src/cow-react/modules/swap/pure/ReceiveAmountInfo/styled.tsx
