+++ src/cow-react/modules/swap/smart/TradeBasicDetails.tsx
@@ -3,7 +3,7 @@ import TradeGp from 'state/swap/TradeGp'
 import { Currency, CurrencyAmount, Percent } from '@uniswap/sdk-core'
 import { INITIAL_ALLOWED_SLIPPAGE_PERCENT } from 'constants/index'
 import { useHigherUSDValue } from 'hooks/useStablecoinPrice'
-import { LowerSectionWrapper } from 'cow-react/modules/swap/dumb/styled'
+import { LowerSectionWrapper } from 'cow-react/modules/swap/pure/styled'
 import { RowFee } from 'components/swap/TradeSummary/RowFee'
 import { RowSlippage } from 'components/swap/TradeSummary/RowSlippage'
 import { RowReceivedAfterSlippage } from 'components/swap/TradeSummary/RowReceivedAfterSlippage'
