+++ src/custom/components/swap/TradeSummary/RowFee.tsx
@@ -9,7 +9,7 @@ import { StyledInfo } from 'pages/Swap/styleds'
 import { RowBetween, RowFixed } from 'components/Row'
 import { MouseoverTooltipContent } from 'components/Tooltip'
 import { AMOUNT_PRECISION, FIAT_PRECISION } from 'constants/index'
-import { LightGreyText } from 'cow-react/modules/swap/dumb/styled'
+import { LightGreyText } from 'cow-react/modules/swap/pure/styled'
 
 export const GASLESS_FEE_TOOLTIP_MSG =
   'On CoW Swap you sign your order (hence no gas costs!). The fees are covering your gas costs already.'
