+++ src/custom/pages/Swap/SwapMod.tsx
@@ -43,15 +43,15 @@ import CowSubsidyModal from 'components/CowSubsidyModal'
 import { AlertWrapper } from './styleds'
 import { setMaxSellTokensAnalytics } from 'utils/analytics'
 import { ImportTokenModal } from 'cow-react/modules/swap/smart/ImportTokenModal'
-import { CompatibilityIssuesWarning } from 'cow-react/modules/swap/dumb/CompatibilityIssuesWarning'
+import { CompatibilityIssuesWarning } from 'cow-react/modules/swap/pure/CompatibilityIssuesWarning'
 import { ConfirmSwapModalSetup, ConfirmSwapModalSetupProps } from 'cow-react/modules/swap/smart/ConfirmSwapModalSetup'
 import { useAtomValue } from 'jotai/utils'
 import { swapConfirmAtom } from 'cow-react/modules/swap/state/swapConfirmAtom'
-import { SwapButton, SwapButtonContext } from 'cow-react/modules/swap/dumb/SwapButton/SwapButton'
+import { SwapButton, SwapButtonContext } from 'cow-react/modules/swap/pure/SwapButton/SwapButton'
 import { RemoveRecipient } from 'cow-react/modules/swap/smart/RemoveRecipient'
-import { Price } from 'cow-react/modules/swap/dumb/Price'
+import { Price } from 'cow-react/modules/swap/pure/Price'
 import { TradeBasicDetails } from 'cow-react/modules/swap/smart/TradeBasicDetails'
-import { BottomGrouping } from 'cow-react/modules/swap/dumb/styled'
+import { BottomGrouping } from 'cow-react/modules/swap/pure/styled'
 import { ArrowWrapperLoader } from 'components/ArrowWrapperLoader'
 import { HighFeeWarning, NoImpactWarning } from 'components/SwapWarnings'
 import { FeesDiscount } from 'cow-react/modules/swap/smart/FeesDiscount'
